Add cli that generates a file with a cwl step
The user is prompted for input on what the cwl step should look like.
At the moment, the template generates a cwl step that accepts a list of
files as input and a directory name as output. For each input file, an
output file is written to the output directory.