Turn javascript part into an angular app
Because angular is much more convenient.

The flask app now serves the html page of the single page application
and all other views return json.

The html page of the application is served by flask, because we want
to be able to start the application from python in a single command.